+++ mjx/mujoco/mjx/_src/io_test.py
@@ -193,22 +193,6 @@ class ModelIOTest(parameterized.TestCase):
           </tendon>
         </mujoco>"""))
 
-  def test_cylinder_not_implemented(self):
-    with self.assertRaises(NotImplementedError):
-      mjx.put_model(mujoco.MjModel.from_xml_string("""
-        <mujoco>
-          <worldbody>
-            <body>
-              <freejoint/>
-              <geom type="cylinder" size="0.05 0.05"/>
-            </body>
-            <body>
-              <freejoint/>
-              <geom size="0.05"/>
-            </body>
-          </worldbody>
-        </mujoco>"""))
-
   def test_margin_gap_mesh_not_implemented(self):
     with self.assertRaises(NotImplementedError):
       mjx.put_model(mujoco.MjModel.from_xml_string("""